Finished refactoring
More like putting stuff in main in other files
